\name{notnaive-package}
\alias{notnaive-package}
\alias{notnaive}
\docType{package}
\title{
\packageTitle{notnaive}
}
\description{
\packageDescription{notnaive}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{notnaive}
\packageIndices{notnaive}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{notnaive}

Maintainer: \packageMaintainer{notnaive}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
